requirements.txt
beautifulsoup4==4.12.2
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.4
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Deprecated==1.2.13
FlightRadarAPI==1.2.7
fonttools==4.41.0
future==0.18.3
gkeepapi==0.14.2
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gpsoauth==1.0.2
greenlet==2.0.2
gspread==4.0.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
keyring==24.2.0
keyrings.alt==4.2.0
kiwisolver==1.4.4
matplotlib==3.7.2
more-itertools==9.1.0
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
Pillow==10.0.0
protobuf==4.23.2
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pygsheets==2.0.6
pyicloud==1.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.7
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.15.0
zipp==3.16.0
